fix: files_outside_copyright_are_identical
